much improved latency on song playback automation initial values working automation recording working property changing undo-ing working undo buffer action collapsing working get time class, and posix driver

+++ trunk/interface/sound_plugin_rack.h
@@ -42,7 +42,7 @@ Q_OBJECT
 	
 public slots:
 	
-	void property_edited_slot(Property *p_prop,float p_old_val);
+	void property_edited_slot(Property *p_prop,double p_old_val);
 	void update_rack();
 	
 public:
